Se agregó interactividad entre el userScreen y los datos de listaUsuarios
